Added more FK (Don't know if it really works, but we are trying xd) I updated some data on the DB/Model to match with each others (boolean type and not byte FE)
Bug : Cannot POST account on the api. (FK problem ? Controller Problem ?)

+++ src/main/java/com/example/demo/model/AccountType.java
@@ -22,7 +22,7 @@ public class AccountType {
     @Column(name="annual_return")
     private double annualReturn;
     @Column(name="age_restriction")
-    private double ageRestriction;
+    private Integer ageRestriction;
     @Column(name="transaction_amount_limit")
     private int transactionAmountLimit;
